Move template literal logic into AST module
