+++ src/ast/domain.ts
@@ -8,7 +8,6 @@ export {
   getStatements,
   isEmpty,
   replaceWithBodyOf,
-  templateElement,
   Primitive,
   forEach
 };
@@ -53,17 +52,6 @@ function isEmpty(statement: t.Statement): boolean {
 
 function replaceWithBodyOf(path: NodePath, node: t.Statement) {
   path.replaceWithMultiple(getStatements(node));
-}
-
-/**
- * Override babel `templateElement()` because it exposes
- * unnecessary implementation details and it's not type-safe.
- */
-function templateElement(value: string): t.TemplateElement {
-  return t.templateElement({
-    raw: value,
-    cooked: value
-  });
 }
 
 type Primitive =
