+++ src/ast/index.ts
@@ -4,4 +4,6 @@ export * from "./scope";
 export * from "./selection";
 export * from "./siblings";
 export * from "./switch";
+export * from "./template-literal";
+export { templateElement } from "./template-literal";
 export * from "./transformation";
